+++ CHANGELOG.md
@@ -10,7 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- In Node.js distribution, use `package.json` engine to configure Babel build
+
 ### Changed
+- Output pure ESM package (no more CommonJS)
 - Update tester
 - Maintenance update